+++ apps/freeablo/falevelgen/levelgen.h
@@ -11,7 +11,7 @@ namespace DiabloExe
 namespace FALevelGen
 {
   
-    FAWorld::GameLevel* generate(size_t width, size_t height, size_t dLvl, const DiabloExe::DiabloExe& exe, size_t previous, size_t next);
+    FAWorld::GameLevel* generate(int32_t width, int32_t height, int32_t dLvl, const DiabloExe::DiabloExe& exe, int32_t previous, int32_t next);
 
 }
